Set up machinery for signature phases, morphisms
The idea is that we can parse and interpret JonPRL signatures in
different phases; then, we can transport signatures along partial
signature-morphisms. So, we'd start with parsing a SIGNATURE[string],
which has the structure of a signature, but all the relevant datatypes
are just strings.

Then we create a signature morphism SIGNATURE[string] >->
SIGNATURE[abt], which actually aparses all the strings into abts, sorts,
etc. We could also have another phase SIGNATURE[valid] w/ morphism
SIGNATURE[abt] >-> SIGNATURE[valid], which actually executes the
included tactic scripts, etc.

The benefit of doing this incrementally rather than all at once is that
we can take advantage of the existing structure whilst doing things like
parsing or executing; this is important, because the nature of our
parser changes as we work our way down a signature (because of notation
declarations, etc.).