Get rid of broken hack where SEQ binds both symbols & variables [details below]
In order to reference a free variable (hypothesis) from within a term
quoted into a tactic script, I had this bizarre hack where the SEQ
tactical would tacitly bind both symbols for hypotheses and variables.
Then in the Nominal LCF elaborator, I would try and rename everything
just so in order that the variables and hypothesis-symbols should
coincide.

This was quite nasty, and it didn't quite work in fact.

I've got rid of this, and added a new operator "@u" ("hyp-ref of u")
which stands for the variable that corresponds to the hypothesis
(symbol) "u"; then, as tactic scripts are elaboratored, *free*
occurrences of hyp-refs are expanded to variables. This is simpler and
clearer, and it makes sure that everything gets expanded at just the
right moment.