parse sequents. breaking change!
The concrete syntax for membership is now 'in', instead of ':'. This is
in order to fix an ambiguity in the grammar between binding a name to a
hypothesized judgement, or hypothesizing a (nameless) membership
judgement.

For consistency, we might want to also change the concrete syntax of the
equality judgment to be `a = b in A`, but I haven't done this yet.