convert abts to sequents.
I don't understand functors very well, so I just banged my head against
it until it compiled. This should be fixed up.

There are at least two issues.

1) Sequents are functorized over the categorical judgments without
   constraints. But in order to convert abts to sequents, we obviously
   need to know that our abts have the operators we expect, and not some
   arbitrary operators. Thus we restrict Tm.O.t to be RedPrlOperator.t.
   I'm pretty confident that something like this is necessary.

2) Are hypothesis names symbols or variables? The current infrastructure
   treats them as symbols (eg, hyp{x}), but they feel like variables in
   the context of the binding structure of sequents (eg, they can appear
   as terms, whereas symbols cannot: b : bool true >> b = b : bool). In
   order to get things to compile, I introduced the completely bogus
   sharing constraint Sym.t = Var.t, but this should be revisited.

As an aside, could we just not functorize sequents over categorical
judgments? I'm not sure exactly what the benefits of the functorization
are, so I can't comment on this further.