don't call nat-rec equality rule in auto
This rule leads to a subgoal (JDG_TERM) which auto can't discharge; this
subgoal is confusing to the user, since they don't know where it came
from. The best solution for now is to make the user manually call this
rule instead.

In the future, it might possibly be cool to come up with a form of judgment like
"figure out the motive of induction" or whatever; then, it might be
semi-automatable, but more importantly, it would be fine if it appeared
in subgoals generated by 'auto', because its meaning would be more clear
to the user.