Remove Record.Elim from AutoElim.
The issue is that `Record.Elim` introduces new variables in the contexts
which do not connect to the original variable. Consider the situation
where we have two goals and the second one

  c : (record ...) >> #1 = (! proj1 c)

depends on the realizer of the first goal. Is it very possible that this
goal is easy when we know what #1 is, but becomes extremely difficult
when we prematurely apply `Record.Elim` on `c`. The reason is that
the new tuple might lose the connection to what #1 will be. Since `auto`
is applied in parallel, we should not apply `Record.Elim` or any rule
that might destroy the connection to other goals eagerly.