Univalence, stage 0. (#338)
* Add the ua operators.
* Add the ua syntax.
* Tweak and refactor toStringPoly.
* Tweak the parsing rules for dimensions.
* Add the parsing rules for ua.
* Add the dynamics of ua.
* Add Univalence.EqType and introduce intoApp.
* Add Univalence.Eq.
* Add Univalence.True.
* Plug in the rules for univalence types.
* Add Synth.Proj and fix Record.EqProj. Fix #343.
* Add TrueHeadExpansion.
* Remove unused intoCoe.
* Unify the registered rule names.
* Tweak the subgoal ordering of Path.EqAppConst.
* Remove Record.Elim from AutoElim.

The issue is that `Record.Elim` introduces new variables in the contexts
which do not connect to the original variable. Consider the situation
where we have two goals and the second one

  c : (record ...) >> #1 = (! proj1 c)

depends on the realizer of the first goal. Is it very possible that this
goal is easy when we know what #1 is, but becomes extremely difficult
when we prematurely apply `Record.Elim` on `c`. The reason is that
the new tuple might lose the connection to what #1 will be. Since `auto`
is applied in parallel, we should not apply `Record.Elim` or any rule
that might destroy the connection to other goals eagerly.

* Add a simple example of ua.
* Update the highlighting.
* Incorporate Jon's Thought.
* (by Jon) fix bug in Synth.Custom; work internally rather than using sequents